pyproject.toml
[project]
name = "egraphics"
requires-python = ">=3.12"

[tool.poetry]
name = "egraphics"
version = "0.18.0"
description = "Graphics Library"
authors = ["Erik Soma <stillusingirc@gmail.com>"]
include = [
    {path = "src/egraphics/*.so", format = "wheel"},
    {path = "src/egraphics/*.pyd", format = "wheel"},
]

[tool.poetry.build]
generate-setup-file = false
script = "buildscript.py"

[tool.poetry.dependencies]
python = "^3.12"
emath = "^0"
egeometry = "^0"
pillow = "^10.4"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-cov = "4.1.0"
eplatform = "^0"
pyopengl = "^3.1.7"

[build-system]
requires = ["poetry-core==1.9.1", "setuptools==69.0.2"]
build-backend = "poetry.core.masonry.api"

[tool.cibuildwheel.linux]
before-all = [
    "yum install mesa-libGL-devel mesa-libGLU-devel -y",
]

[[tool.cibuildwheel.overrides]]
select = "*-musllinux*"
before-all = "apk add mesa-gl glu-dev"

[virtualenvs]
in-project = true

[tool.pyright]
venvPath = "."
venv = ".venv"
